Divide the MuFi digi signal on the number of readout sides.
This is the simplest way to distribute the deposited energy. In the future we aim to provide
a more sophisticated model taking into account light reflection and its effect on
the number of photons seen by a SiPM. This effect could be significant when the particles
pass close to the readout edge.