Revert commit b7c521656 to bring the special vetoPoint back
These points are actually used in the Floor class that describes
the tunnel and the upstream rock and are a handle to study
interactions in those insensitive volumes.
For example, vetoPoint objects are used in a special case of
muonDIS simulations to study the energy loss in the upstream rock.
It is important to always have a branch for them, even if empty most
of the time, as one can use different MC samples together: one with
no MC point information about interactions in the rock and one
where this is included.