sndLHC_TI18geom_config: set all spatial- and time-calibration constants to zero
This is conveninet as one will be using this geo configuration for MC and
as a base for future testbeams' geo configs. For testbeams, the alignment
can be added straight to the geo config as we currently do. In MC the
detector is perfectly positioned and time-aligned by default.
After this commit, all calibration constants are to be added together when
the geofile is modified.