ShowerTools: add Scifi and US spatial anisotropy
The functions take as arguments Scifi/US plane hits. There is an option to use plane
centroids for the anisotropy estimation - to be tested if makes sense.
It is up to the user to filter in time and/or find the centroids(if flag is set to 1)
prior the call to anisotropy functions. This is done to remove redundancies and provide
flexibility to the users.